+++ src/yabf/core/__init__.py
@@ -1,3 +1,4 @@
 """Core functionality for YABF."""
 
 from . import _samplers, component, configio, io, likelihood, mpi, parameters, utils
+from ._samplers import Sampler
